update to the KGX format used by the text-mined KG
The main change is the addition of an evidence node to store the NLP-related metadata that supports the assertion of the Biolink association represented by an edge. The evidence node schema contains columns for the sentence, publication CURIE, the spans for the subject, relation, and object that were mined from the sentence, and a reference to who/what provided the mined assertion. Each edge references one or more evidence nodes by a pipe-delimited list of evidence node identifiers.

Changes made are based on this discussion: https://github.com/NCATS-Tangerine/kgx/issues/174